pyproject.toml
[project]
name = "ingestor"
version = "0.1.0"
description = "Comprehensive media-to-markdown ingestion for LLM RAG and fine-tuning"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Jaime Cernuda" }
]
keywords = ["markdown", "ingestion", "rag", "llm", "document-conversion", "crawl4ai", "whisper"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI & Output
    "click>=8.1.0",
    "rich>=13.0.0",
    # Config
    "pyyaml>=6.0",
    "pydantic>=2.0",
    # Core Infrastructure
    "magika>=0.5.0", # AI file type detection (Google, 99% accuracy)
    "charset_normalizer>=3.0.0", # Charset detection for non-UTF8
    "markdownify>=0.11.0", # HTML -> Markdown
    # Images (always needed for extraction)
    "pillow>=10.0.0",
    # Async & HTTP
    "aiofiles>=23.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    "crawl4ai>=0.7.8",
    "mammoth>=1.11.0",
    "google-genai>=1.56.0",
]

[project.optional-dependencies]
# Document formats
docx = ["docx2python>=2.0.0", "mammoth>=1.6.0"]  # python-docx only for fixtures (in dev)
pptx = ["python-pptx>=0.6.21"]
epub = ["ebooklib>=0.18"]
pdf = ["docling>=2.0.0", "pymupdf>=1.26.6"]  # Docling for ML extraction, PyMuPDF for fallback

# Spreadsheets
xlsx = ["pandas>=2.0.0", "openpyxl>=3.1.0"]
xls = ["pandas>=2.0.0", "xlrd>=2.0.0"]

# Web
web = ["crawl4ai>=0.7.8"]
youtube = ["youtube-transcript-api>=1.2.0", "yt-dlp>=2024.0.0"]
git = []  # Uses httpx from core dependencies for GitHub API + subprocess git

# Audio
audio = ["openai-whisper>=20231117"]

# Data formats
xml = ["defusedxml>=0.7.0"]
csv = ["pandas>=2.0.0"]

# Scientific papers (DOI, arXiv, etc.)
paper = [
    "ingestor[pdf]",  # PDF extraction for downloaded papers
]

# AI features (optional)
vlm = ["ollama>=0.2.0"]
agent = ["anthropic>=0.39.0"]  # For Claude API integration

# Bundles
all-formats = [
    "ingestor[docx,pptx,epub,xlsx,xls,web,youtube,git,audio,xml,csv,pdf,paper]"
]
all = [
    "ingestor[all-formats,vlm,agent]"
]

dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    # For test fixture generation
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
    "ebooklib>=0.18",
]

[project.scripts]
ingestor = "ingestor.cli:main"
researcher = "researcher.cli:main"
parser = "parser.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingestor", "src/researcher", "src/parser"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
# Disable strict mode to reduce noise during incremental adoption
strict = false
# Enable useful checks without being too strict
warn_return_any = false
warn_unused_ignores = true
check_untyped_defs = true
# These are too noisy for gradual typing adoption
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false

# Ignore missing stubs for third-party libraries
[[tool.mypy.overrides]]
module = ["yaml", "aiofiles"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "network: requires network access (run with --network)",
    "skip_audio: skip audio tests (can skip with -m 'not skip_audio')",
    "slow: marks tests as slow running",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
